test: unit test detectUsedComponents directly
